requirements.txt
streamlit==1.22.0
pandas
Pillow
numpy